backend/requirements.txt
# Core AI Framework
crewai>=0.193.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20

# API Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.6

# HTTP Client
httpx>=0.25.1

